reorganized imageseries, moving ImageSeries to baseclass module; loading process from __init__, and renaming adapters to load, write to save
