+++ hexrd/imageseries/imageseriesabc.py
@@ -3,4 +3,3 @@ import collections
 
 class ImageSeriesABC(collections.Sequence):
     pass
-    # define interface here
