+++ hexrd/imageseries/load/array.py
