+++ hexrd/imageseries/load/framecache.py
