+++ hexrd/imageseries/load/registry.py
