+++ hexrd/imageseries/load/trivial.py
