+++ hexrd/imageseries/save.py
